Updated to Tabulator version 4.9.3 (from 4.9.1)

+++ esasky-cl/src/main/webapp/js/tabulator/tabulator.js
@@ -1,6 +1,6 @@
 var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol" ? function (obj) { return typeof obj; } : function (obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; };
 
-/* Tabulator v4.9.1 (c) Oliver Folkerd */
+/* Tabulator v4.9.3 (c) Oliver Folkerd */
 
 ;(function (global, factory) {
 	if ((typeof exports === 'undefined' ? 'undefined' : _typeof(exports)) === 'object' && typeof module !== 'undefined') {
@@ -2350,25 +2350,30 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
 			var definition;
 
 			if (!_this10.isGroup) {
-				definition = Object.assign({}, _this10.getDefinition());
-				definition = Object.assign(definition, updates);
+				if (!_this10.parent.isGroup) {
+					definition = Object.assign({}, _this10.getDefinition());
+					definition = Object.assign(definition, updates);
 
-				_this10.table.columnManager.addColumn(definition, false, _this10).then(function (column) {
+					_this10.table.columnManager.addColumn(definition, false, _this10).then(function (column) {
 
-					if (definition.field == _this10.field) {
-						_this10.field = false; //cleair field name to prevent deletion of duplicate column from arrays
-					}
+						if (definition.field == _this10.field) {
+							_this10.field = false; //cleair field name to prevent deletion of duplicate column from arrays
+						}
 
-					_this10.delete().then(function () {
-						resolve(column.getComponent());
+						_this10.delete().then(function () {
+							resolve(column.getComponent());
+						}).catch(function (err) {
+							reject(err);
+						});
 					}).catch(function (err) {
 						reject(err);
 					});
-				}).catch(function (err) {
-					reject(err);
-				});
+				} else {
+					console.warn("Column Update Error - The updateDefinition function is only available on ungrouped columns");
+					reject("Column Update Error - The updateDefinition function is only available on columns, not column groups");
+				}
 			} else {
-				console.warn("Column Update Error - The updateDefinition function is only available on columns, not column groups");
+				console.warn("Column Update Error - The updateDefinition function is only available on ungrouped columns");
 				reject("Column Update Error - The updateDefinition function is only available on columns, not column groups");
 			}
 		});
@@ -4158,7 +4163,8 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
 		    modExists;
 
 		if (this.renderMode === "virtual") {
-			var otherHeight = Math.floor(this.columnManager.getElement().getBoundingClientRect().height + (this.table.footerManager && !this.table.footerManager.external ? this.table.footerManager.getElement().getBoundingClientRect().height : 0));
+
+			var otherHeight = Math.floor(this.columnManager.getElement().getBoundingClientRect().height + (this.table.footerManager && this.table.footerManager.active && !this.table.footerManager.external ? this.table.footerManager.getElement().getBoundingClientRect().height : 0));
 
 			if (this.fixedHeight) {
 				this.element.style.minHeight = "calc(100% - " + otherHeight + "px)";
@@ -4696,7 +4702,7 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
 			for (var _i7 = this.leftCol; _i7 <= this.rightCol; _i7++) {
 				var column = this.columns[_i7];
 
-				if (column.visible) {
+				if (column && column.visible) {
 					var cell = row.getCell(column);
 
 					if (cell !== undefined) {
@@ -11039,14 +11045,14 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
 		return output;
 	};
 
-	DataTree.prototype.getChildren = function (row) {
+	DataTree.prototype.getChildren = function (row, allChildren) {
 		var _this48 = this;
 
 		var config = row.modules.dataTree,
 		    children = [],
 		    output = [];
 
-		if (config.children !== false && config.open) {
+		if (config.children !== false && (config.open || allChildren)) {
 			if (!Array.isArray(config.children)) {
 				config.children = this.generateChildren(row);
 			}
@@ -11935,7 +11941,13 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
 				rightEdge -= parseInt(this.table.modules.frozenColumns.rightMargin);
 			}
 
+			if (this.table.options.virtualDomHoz) {
+				leftEdge -= parseInt(this.table.vdomHoz.vDomPadLeft);
+				rightEdge -= parseInt(this.table.vdomHoz.vDomPadLeft);
+			}
+
 			if (cellEl.offsetLeft < leftEdge) {
+
 				this.table.rowManager.element.scrollLeft -= leftEdge - cellEl.offsetLeft;
 			} else {
 				if (cellEl.offsetLeft + cellEl.offsetWidth > rightEdge) {
@@ -12924,7 +12936,7 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
 					success(item.value);
 				}
 
-				initialDisplayValue = input.value;
+				initialDisplayValue = [item.value];
 			}
 
 			function chooseItems(silent) {
@@ -12938,7 +12950,7 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
 					output.push(item.value);
 				});
 
-				initialDisplayValue = input.value;
+				initialDisplayValue = output;
 
 				success(output);
 			}
@@ -13468,6 +13480,8 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
 
 			function showList() {
 				if (!listEl.parentNode) {
+
+					console.log("show", initialDisplayValue);
 					while (listEl.firstChild) {
 						listEl.removeChild(listEl.firstChild);
 					}var offset = Tabulator.prototype.helpers.elOffset(cellEl);
@@ -14764,7 +14778,7 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
 
 			//set empty value function
 			column.modules.filter.emptyFunc = column.definition.headerFilterEmptyCheck || function (value) {
-				return !value && value !== "0";
+				return !value && value !== "0" && value !== 0;
 			};
 
 			filterElement = document.createElement("div");
@@ -15204,7 +15218,9 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
 		self.prevHeaderFilterChangeCheck = "{}";
 
 		this.headerFilterColumns.forEach(function (column) {
-			column.modules.filter.value = null;
+			if (typeof column.modules.filter.value !== "undefined") {
+				delete column.modules.filter.value;
+			}
 			column.modules.filter.prevSuccess = undefined;
 			self.reloadHeaderFilter(column);
 		});
@@ -19578,8 +19594,6 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
 		var receiver = false,
 		    success = false;
 
-		console.trace("drop");
-
 		e.stopImmediatePropagation();
 
 		switch (_typeof(this.table.options.movableRowsReceiver)) {
@@ -20098,7 +20112,7 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
 
 			page = parseInt(page);
 
-			if (page > 0 && page <= _this76.max) {
+			if (page > 0 && page <= _this76.max || _this76.mode !== "local") {
 				_this76.page = page;
 				_this76.trigger().then(function () {
 					resolve();
@@ -20619,7 +20633,7 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
 		}
 
 		if (this.config.columns) {
-			this.load("columns", this.table.options.columns);
+			this.table.options.columns = this.load("columns", this.table.options.columns);
 		}
 	};
 
@@ -22406,7 +22420,7 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
 	};
 
 	SelectRow.prototype.childRowSelection = function (row, select) {
-		var children = this.table.modules.dataTree.getChildren(row);
+		var children = this.table.modules.dataTree.getChildren(row, true);
 
 		if (select) {
 			for (var _iterator2 = children, _isArray2 = Array.isArray(_iterator2), _i19 = 0, _iterator2 = _isArray2 ? _iterator2 : _iterator2[Symbol.iterator]();;) {
@@ -22775,12 +22789,8 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
 		a = typeof a !== "undefined" ? a : "";
 		b = typeof b !== "undefined" ? b : "";
 
-		//ESASky patch
-		if (column.field == "rowSelection") {
-			el1Comp = el1.getComponent();
-			el2Comp = el2.getComponent();
-		}
-		//End ESASky patch
+		el1Comp = el1.getComponent();
+		el2Comp = el2.getComponent();
 
 		return column.modules.sort.sorter.call(this, a, b, el1Comp, el2Comp, column.getComponent(), dir, params);
 	};
